bugfix: rename path, added type button
