+++ src/components/button/Button.tsx
@@ -6,14 +6,18 @@ interface Props {
   withBorder: boolean;
   children?: ReactNode;
   absoluteParms?: string;
+  type: 'button' | 'submit';
 }
 
-export const Button = ({ name, bgColor, withBorder, children, absoluteParms }: Props) => {
+export const Button = ({ name, bgColor, withBorder, children, absoluteParms, type }: Props) => {
   const background = bgColor ? bgColor : 'bg-inherit ';
   const border = withBorder ? 'border border-zinc-200 hover:border-zinc-300' : 'border-none';
   const absolute = absoluteParms ? absoluteParms : 'static ';
   return (
-    <button className={absolute + border + ' text-zinc-100 font-bold py-1.5 px-2 hover:text-zinc-300 ' + background}>
+    <button
+      type={type}
+      className={absolute + border + ' text-zinc-100 font-bold py-1.5 px-2 hover:text-zinc-300 ' + background}
+    >
       {name}
       {children}
     </button>
