+++ src/components/characterCard/CharacterCard.tsx
@@ -1,0 +1,3 @@
+export const CharacterCard = () => {
+  return <div></div>;
+};
