+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './menu/Menu';
 export * from './searchBar/SearchBar';
 export * from './button/Button';
+export * from './characterCard/CharacterCard';
