+++ src/components/menu/Menu.tsx
@@ -2,8 +2,7 @@ import { Link } from 'react-router-dom';
 
 import logo from '../../assets/image/logo.png';
 import { ReactComponent as Sun } from '../../assets/svg/sun.svg';
-import { Button } from '../button/Button';
-import { SearchBar } from '../searchBar/SearchBar';
+import { Button, SearchBar } from '../../components';
 
 export const Menu = () => {
   // TODO: Убрать когда будет приходить через пропсы
@@ -20,27 +19,27 @@ export const Menu = () => {
       </div>
       <div className="border-zinc-100/20 border-l">
         <div className="flex gap-4 px-4 items-center ">
-          <Button withBorder={false}>
+          <Button type={'button'} withBorder={false}>
             <Sun fill="#e4e4e7" stroke="#e4e4e7"></Sun>
           </Button>
 
           {isAuth ? (
             <>
               <Link className="text-zinc-200 font-bold py-1 px-2" to="/favorites">
-                <Button name={'Favorites'} withBorder={false} />
+                <Button type={'button'} name={'Favorites'} withBorder={false} />
               </Link>
               <Link className="text-zinc-200 font-bold py-1 px-2" to="/history">
-                <Button name={'History'} withBorder={false} />
+                <Button type={'button'} name={'History'} withBorder={false} />
               </Link>
-              <Button name={'Exit'} withBorder={true} />
+              <Button type={'button'} name={'Exit'} withBorder={true} />
             </>
           ) : (
             <>
               <Link to="/signin">
-                <Button name={'Sign In'} withBorder={false} />
+                <Button type={'button'} name={'Sign In'} withBorder={false} />
               </Link>
               <Link to="/signup">
-                <Button name={'Sign Up'} withBorder={true} />
+                <Button type={'button'} name={'Sign Up'} withBorder={true} />
               </Link>
             </>
           )}
